requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.9.2
asn1crypto==1.5.1
billiard==4.2.2
braintree==4.39.0
Brotli==1.1.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.2
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.7
django-braces==1.17.0
django-embed-video==1.4.10
django-extensions==4.1
django-localflavor==5.0
django-memcache-status==2.3
django-parler==2.3
django-rosetta==0.10.2
django-taggit==6.1.0
djangorestframework==3.16.1
dotenv==0.9.9
easy-thumbnails==2.10.1
fonttools==4.60.1
freetype-py==2.5.1
html5lib==1.1
idna==3.10
kombu==5.5.4
logger==1.4
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.3
oauthlib==3.3.1
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
polib==1.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==6.1.2
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-memcached==1.62
python-stdnum==2.1
python3-openid==3.2.0
PyYAML==6.0.3
redis==7.0.0
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rlPyCairo==0.4.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.8.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
